moved WorkplaceCategory and WorkplaceService within same model file
