--- src/app/core/model/workplace-category.model.ts
@@ -1,6 +1,0 @@
-import { WorkplaceService } from '@core/model/workplace-service.model';
-
-export interface WorkplaceCategory {
-  category: string;
-  services: Array<WorkplaceService>;
-}
